backend/pyproject.toml
[project]
name = "colbert-backend"
version = "0.1.0"
description = "RAG-powered chatbot for French public administration information"
authors = [
    { name = "Your Name", email = "your.email@example.com" },
]
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "python-dotenv>=0.19.0",
    "langchain>=0.1.0",
    "langchain-core>=0.1.0",
    "langchain-mistralai>=0.0.1",
    "loguru>=0.7.0",
    "redis>=4.5.0",
    "mistralai>=1.7.0",
    "langchain-redis>=0.2.0",
    "langchain-tavily>=0.1.6",
    "chromadb>=1.0.9",
    "langchain-community>=0.3.23",
    "ipykernel>=6.29.5",
    "langchain-chroma>=0.2.4",
]
requires-python = ">=3.11"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
